Remove the disabled method for setting a custom protocol
Firstly, the current QCoapProtocol interface only supports customizing
some parameters like timeouts, counters, etc., which can also be done
via the QCoapClient class, so there is no need for a setter method for
the protocol.

Secondly, the protocol runs on a separate thread, so it is not a good
idea to allow users to change the protocol at runtime. Even if later we
decide to support custom protocols, it can be set during the
construction of the client. So the setter method is not needed.

Removed the setter method and added the missing methods for customizing
protocol-related parameters.

Change-Id: I89375ac502be4d29d3cb617da9c2f99a3ab410ec
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>