Hide the APIs of QCoapConnection and QCoapQUdpConnection
At the moment only the QCoapQUdpConnection class implements the
QCoapConnection interface. If later we decide to add other
implementations, and it turns out that the current API is not good
enough, it won't be possible to change it  without breaking the source
compatibility. Let's make these classes private, we can make them public
when there is a need for it.

Change-Id: I41336dda0e6eaa762f0eb9c8f4aa98a9c7b62a2c
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>