conanfile.py: Use QTDIR instead of QT_PATH
There's an established environment variable for this though, QTDIR.
This environment variable is also set by Qt Creator by default.
So the Conan recipes should accept this environment variable instead.

This requires a change in the matching conan profile files
where the path to Qt installation is injected to env.

Pick-to: 6.1
Task-number: QTBUG-90583
Change-Id: I33e2ebc6b04f87027e039d0ae951179f5b2d5530
Reviewed-by: Antti Kokko <antti.kokko@qt.io>