Conan: Inherit recipe class from QtLeafModule for common functionality
The qt-conan-common package implements a base class for Qt leaf
module recipes. The build steps in leaf modules are mostly identical
so it makes sense to put those in the base class.

Dependencies are read by the base class from the 'dependencies.yaml'
which is the same file the CI system uses.

Pick-to: 6.2
Task-number: QTBUG-94867
Change-Id: I8b80055e468374ebda3823dbf3042a1855d3cf40
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Toni Saario <toni.saario@qt.io>