Doc: Repairs broken \externalpage URLs and update redirected ones
`src/coap/doc/src/external-resources.qdoc` defines a series of
`\externalpage` resources to some RFCs.

The links to `https://tools.ietf.org/search/*` resolve to an `Object
not found!` page.

To avoid the problem, they were replaced with an equivalent link on
the `datatracker.ietf.org` domain, which is one of the official ways
of reading published RFCs as per the IETF website.

The `tools.ietf.org/html/*` web pages containing RFCs now redirect
to `datatracker.ietf.org/doc/html/*`.

While the redirection work, our script to catch broken links on
dev-snapshots builds of the documentation doesn't handle redirection
correctly, reporting it as broken.

To appease the broken-links script, to avoid an unneccesary and to be
consistent in the domain that is linked to by the other pages, the
relevant `\externalpage` link were modified to point to the new domain.

Task-number: QTBUG-96127
Change-Id: Idcaed482b1d54eefe33cb157fca425b2abd47962
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 3454d3e73428b71e9989556447532ac0d38ae9cc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>