Update commercial license headers
Updated header.COMM to the files in tqtc-qtcoap..
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4927
Change-Id: I2e61c9c859e088da0ae33cc695261a479599035f
Reviewed-by: Sona Kurazyan <sona.kurazyan@qt.io>