QtCoap: replace qSwap with member-swap where possible
qSwap() is a monster that looks for ADL overloads of swap() and also
detects the noexcept of the wrapped swap() function, so it should only
be used when the argument type is unknown. In the vast majority of
cases, the type is known to be efficiently std::swap()able or to have
a member-swap. Call either of these.

Task-number: QTBUG-97601
Pick-to: 6.3 6.2 5.15
Change-Id: Icaed67ff42bc148cb70e42d3377dea6b7eed908e
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>