Simple CoAP Client: extend options dialog
Add possibility to remove individual entries from the options table.
Also change the selection behavior of the table to allow selection of
individual rows only. This makes sense, because we do not allow to
change values in the already-added entries anyway, and row-selection
makes more sense when deleting entries.

Task-number: QTBUG-113858
Pick-to: 6.6 6.5
Change-Id: Ica56c697c79c758d473f647831afc793c7ca35b0
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>