Simple CoAP Client: remember options in options dialog
The options dialog is re-created every time we press the "Add Options"
button. As a result, it looses all the options that were set, even
if the message was not yet sent. This is inconvenient, as it does not
allow the user to append forgotten options, or simply to double-check
which options are used.

This patch passes the current options to the OptionDialog's c-tor, so
that the dialog opens with the list of the relevant options.

Task-number: QTBUG-113858
Pick-to: 6.6 6.5
Change-Id: Ia11754706884229b3971526e94c0f8470b5d89b2
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>