CoAP Multicast Discovery example: fix error handling
The CoapMulticastClient.onError handler was actually trying to handle
the base QCoapClient::error signal which is using a QCoapReply* as a
parameter. This type is not known to QML, so this produced a qmllint
warning.

Fix it by processing the QCoapClient::error() signal on the C++ side,
and converting it to the existing QmlCoapMulticastClient::finished()
signal with only one int parameter, because the QCoapReply is really
not needed for this example.

Task-number: QTBUG-113858
Change-Id: I5b41dd9d19f44d813051f4705521d0026ab55848
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit ebd40ab7302e36a58596a486bc68a2587ad1eeae)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>