CoAP Multicast Discovery example: fix resource delegate height
The height of the delegate which shows the discovered resources was
hardcoded to 60 px. It works fine until someone wants to increase the
default system font size, in which case the data will not fit into the
rectangle anymore.

Fix it by using the implicitHeight of the Column which actually
contains the data. As a drive-by - also add a bottom padding, to center
the content in the delegate rectangle.

Task-number: QTBUG-113858
Pick-to: 6.6 6.5
Change-Id: I25349005d2edf236ff9e685d7cd8519e30625f97
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>