Docs: use \keyword on QtCoap namespace
The QtCoap namespace clashes with the QtCoap module, when it comes to
linking in qdoc. Because of that it was not possible to link to the
namespace.
Add a

 \keyword {QtCoap Namespace}

command to the QtCoap namespace, so that it could be
referenced using \l {QtCoap Namespace}.

Change-Id: I066504dac6107a5bbc7f5d40a7f0c3924c2e8acb
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
(cherry picked from commit 355d0ea260823dbb4d30db397325c2518600249d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>