QCoapReply: add test for a chunked read
QCoapReply is a QIODevice, but it was not obvious if readData() is
implemented properly, or if it also needed to update pos().

This patch adds a unit test that checks a chunked read from the reply.
The test passes, so nothing needs to be fixed in the implementation.

Pick-to: 6.9 6.8 6.5
Change-Id: I0642718f21a31cf978a733bc3bc742ad1eb47b3a
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit f310b4a5a296cc4ec8d3a3d81856cb9e4d74bd6a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>