QCoapResourceDiscoveryReply: store QString's size in qsizetype
... instead of an int. The int variable had a risk of overflow, which
is technically an UB. However, in practice it would have silently
resulted in a negative value.
The only usage of the variable is a call to QString::mid(), which
handles the negative second parameter as "provide data until the end
of the string".

Pick-to: 6.9 6.8 6.5
Change-Id: Ic95de0ceba09e02b726ab90407d637c449da2ccd
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit d85ded9e4029025c911f0f88bd7a2de2477fd683)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>